Bugfix
Using a dictionary to story the departures, indexed by departure
time `datetime` object was a mistake, because two `datetime`
objects with the same time have the same hash, causing departures
with timestamp collission to be dropped from the dictionary.

Changed to using a list of (datetime, departure) tuples, which
sorts more or less correctly and preserves uniqueness.